rearranged types
