+++ packages/react/src/lib/positioner/index.ts
@@ -1,2 +1,2 @@
 export * from './types'
-export { default as Popper, PositionCommonProps, PopperChildrenProps } from './Popper'
+export { default as Popper } from './Popper'
